Add filter tag

+++ user/views.py
@@ -7,25 +7,12 @@ from django.conf import settings
 from django.core import serializers
 from django.http import HttpResponse
 from django.shortcuts import render, redirect
-from django.core.files.storage import FileSystemStorage
 from django.contrib.auth.mixins import LoginRequiredMixin
 from qr_code.qrcode.utils import WifiConfig, QRCodeOptions
 
 from user.models import User, Document
 from .forms import NameForm, DocumentForm
-
-
-class FileUpload():
-
-	def upload(path, file):
-		file_path = path+'/'+datetime.date.today().isoformat()
-		fs = FileSystemStorage(location='media/'+file_path)
-		tempname = str(time.time_ns())+'.png'
-		filename = fs.save(tempname, file)
-		file_original_path = file_path+'/'+str(filename)
-		uploaded_file_url = fs.url(file_original_path)
-
-		return uploaded_file_url
+from applibs.file_upload import FileUpload
 
 
 # Create your views here.
@@ -64,18 +51,24 @@ class UserView(View):
 
 		# check whether it's valid:
 		if form.is_valid():
-			email = request.POST.get('email')
-			password = request.POST.get('password')
-			first_name = request.POST.get('first_name')
-			last_name = request.POST.get('last_name')
 
-			if request.FILES.get('profile_img'):
-				file = request.FILES.get('profile_img')
-				uploaded_file_url = FileUpload.upload('users', file)
-				# print(uploaded_file_url)
+			try:
+				email = request.POST.get('email')
+				password = request.POST.get('password')
+				first_name = request.POST.get('first_name')
+				last_name = request.POST.get('last_name')
 
-			user = User.objects.create_user(email, password, first_name=first_name, last_name=last_name)
-			document = Document.objects.save(user, uploaded_file_url)
+				user = User.objects.create_user(email, password, first_name=first_name, last_name=last_name)
+
+				if request.FILES.get('profile_img'):
+					file = request.FILES.get('profile_img')
+					uploaded_file_url = FileUpload.upload('users', file)
+					# print(uploaded_file_url)
+					document = Document.objects.save(user, uploaded_file_url)
+
+			except Exception as e:
+				raise e
+
 			return redirect('view_user')
 
 		return redirect('new_user')
@@ -121,22 +114,24 @@ class UserEdit(View):
 
 		# check whether it's valid
 		if form.is_valid():
-			user = User.objects.update_user(id, request.POST)
-			old_file = Document.objects.get_document(user_id=id)
 
 			try:
-				delete_file = settings.BASE_DIR+old_file.document
+				user = User.objects.update_user(id, request.POST)
+				old_file = Document.objects.get_document(user_id=id)
 
 				if request.FILES.get('profile_img'):
-					if os.path.isfile(delete_file):
-						os.remove(delete_file)
+
+					if old_file:
+						delete_file = settings.BASE_DIR+old_file.document
+						if os.path.isfile(delete_file):
+							os.remove(delete_file)
+						Document.objects.delete(id=old_file.id)
 
 					file = request.FILES.get('profile_img')
 					uploaded_file_url = FileUpload.upload('users', file)
 					# print(uploaded_file_url)
-
-					Document.objects.delete(id=old_file.id)
 					document = Document.objects.save(user, uploaded_file_url)
+
 			except Exception as e:
 				raise e
 
@@ -150,12 +145,13 @@ class UserDelete(View):
 	def get(self, request, id):
 
 		try:
-			old_file = Document.objects.get_document(user_id=id)
-			delete_file = settings.BASE_DIR+old_file.document
 			User.objects.delete_user(id)
+			old_file = Document.objects.get_document(user_id=id)
 
-			if os.path.isfile(delete_file):
-				os.remove(delete_file)
+			if old_file:
+				delete_file = settings.BASE_DIR+old_file.document
+				if os.path.isfile(delete_file):
+					os.remove(delete_file)
 
 			return True
 		except Exception as e:
